# Return optional value in to_json functions (fixes #1064)