--- a/boa/src/builtins/json/mod.rs
+++ b/boa/src/builtins/json/mod.rs
@@ -178,10 +178,11 @@ impl Json {
         let replacer = match args.get(1) {
             Some(replacer) if replacer.is_object() => replacer,
             _ => {
-                return Ok(Value::from(json_to_pretty_string(
-                    &object.to_json(context)?,
-                    gap,
-                )))
+                if let Some(value) = object.to_json(context)? {
+                    return Ok(Value::from(json_to_pretty_string(&value, gap)));
+                } else {
+                    return Ok(Value::undefined());
+                }
             }
         };
 
@@ -208,10 +209,11 @@ impl Json {
                             ),
                         );
                     }
-                    Ok(Value::from(json_to_pretty_string(
-                        &object_to_return.to_json(context)?,
-                        gap,
-                    )))
+                    if let Some(value) = object_to_return.to_json(context)? {
+                        Ok(Value::from(json_to_pretty_string(&value, gap)))
+                    } else {
+                        Ok(Value::undefined())
+                    }
                 })
                 .ok_or_else(Value::undefined)?
         } else if replacer_as_object.is_array() {
@@ -234,8 +236,9 @@ impl Json {
             for field in fields {
                 let v = object.get_field(field.to_string(context)?, context)?;
                 if !v.is_undefined() {
-                    let value = v.to_json(context)?;
-                    obj_to_return.insert(field.to_string(context)?.to_string(), value);
+                    if let Some(value) = v.to_json(context)? {
+                        obj_to_return.insert(field.to_string(context)?.to_string(), value);
+                    }
                 }
             }
             Ok(Value::from(json_to_pretty_string(
@@ -243,10 +246,11 @@ impl Json {
                 gap,
             )))
         } else {
-            Ok(Value::from(json_to_pretty_string(
-                &object.to_json(context)?,
-                gap,
-            )))
+            if let Some(value) = object.to_json(context)? {
+                Ok(Value::from(json_to_pretty_string(&value, gap)))
+            } else {
+                Ok(Value::undefined())
+            }
         }
     }
 }