--- a/boa/src/value/mod.rs
+++ b/boa/src/value/mod.rs
@@ -215,7 +215,7 @@ impl Value {
     }
 
     /// Converts the `Value` to `JSON`.
-    pub fn to_json(&self, context: &mut Context) -> Result<JSONValue> {
+    pub fn to_json(&self, context: &mut Context) -> Result<Option<JSONValue>> {
         let to_json = self.get_field("toJSON", context)?;
         if to_json.is_function() {
             let json_value = context.call(&to_json, self, &[])?;
@@ -223,27 +223,25 @@ impl Value {
         }
 
         match *self {
-            Self::Null => Ok(JSONValue::Null),
-            Self::Boolean(b) => Ok(JSONValue::Bool(b)),
+            Self::Null => Ok(Some(JSONValue::Null)),
+            Self::Boolean(b) => Ok(Some(JSONValue::Bool(b))),
             Self::Object(ref obj) => obj.to_json(context),
-            Self::String(ref str) => Ok(JSONValue::String(str.to_string())),
+            Self::String(ref str) => Ok(Some(JSONValue::String(str.to_string()))),
             Self::Rational(num) => {
                 if num.is_finite() {
-                    Ok(JSONValue::Number(
+                    Ok(Some(JSONValue::Number(
                         JSONNumber::from_str(&Number::to_native_string(num))
                             .expect("invalid number found"),
-                    ))
+                    )))
                 } else {
-                    Ok(JSONValue::Null)
+                    Ok(Some(JSONValue::Null))
                 }
             }
-            Self::Integer(val) => Ok(JSONValue::Number(JSONNumber::from(val))),
+            Self::Integer(val) => Ok(Some(JSONValue::Number(JSONNumber::from(val)))),
             Self::BigInt(_) => {
                 Err(context.construct_type_error("BigInt value can't be serialized in JSON"))
             }
-            Self::Symbol(_) | Self::Undefined => {
-                unreachable!("Symbols and Undefined JSON Values depend on parent type");
-            }
+            Self::Symbol(_) | Self::Undefined => Ok(None),
         }
     }