<!---
Thank you for contributing to Boa! Please fill out the template below, and remove or add any
information as you feel neccesary.
--->

This Pull Request fixes/closes #1064.

It changes the following:

- Makes `Value::to_json`, `GcObject::to_json` return an optional `JSONValue`. If the value cannot be converted into a JSON value, the functions return `None`.